avg-review is now working for all entities

+++ services/src/main/java/ar/edu/itba/paw/services/ReviewServiceImpl.java
@@ -1,10 +1,16 @@
 package ar.edu.itba.paw.services;
 
+import ar.edu.itba.paw.models.Album;
+import ar.edu.itba.paw.models.Artist;
+import ar.edu.itba.paw.models.Song;
 import ar.edu.itba.paw.models.reviews.Review;
 import ar.edu.itba.paw.models.reviews.ArtistReview;
 import ar.edu.itba.paw.models.reviews.AlbumReview;
 import ar.edu.itba.paw.models.reviews.SongReview;
+import ar.edu.itba.paw.persistence.AlbumDao;
+import ar.edu.itba.paw.persistence.ArtistDao;
 import ar.edu.itba.paw.persistence.ReviewDao;
+import ar.edu.itba.paw.persistence.SongDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,10 +23,33 @@ import java.util.Optional;
 public class ReviewServiceImpl implements ReviewService {
 
     private final ReviewDao reviewDao;
+    private final SongDao songDao;
+    private final ArtistDao artistDao;
+    private final AlbumDao albumDao;
 
     @Autowired
-    public ReviewServiceImpl(ReviewDao reviewDao) {
+    public ReviewServiceImpl(ReviewDao reviewDao, SongDao songDao, ArtistDao artistDao, AlbumDao albumDao) {
         this.reviewDao = reviewDao;
+        this.songDao = songDao;
+        this.artistDao = artistDao;
+        this.albumDao = albumDao;
+    }
+
+    @Override
+    public int updateAvgRatingForAll(){
+        List<Song> songs = songDao.findAll();
+        List<Album> albums = albumDao.findAll();
+        List<Artist> artists = artistDao.findAll();
+        for (Song song : songs) {
+            updateSongRating(song.getId());
+        }
+        for (Album album : albums) {
+            updateAlbumRating(album.getId());
+        }
+        for (Artist artist : artists) {
+            updateArtistRating(artist.getId());
+        }
+        return 0;
     }
 
     @Override
@@ -30,7 +59,20 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public int deleteById(long id) {
-        return reviewDao.deleteById(id);
+        int res = 0;
+        if (isArtistReview(id)) {
+            res = reviewDao.deleteById(id);
+            updateArtistRating(id);
+        }
+        if (isAlbumReview(id)) {
+            res = reviewDao.deleteById(id);
+            updateAlbumRating(id);
+        }
+        if (isSongReview(id)) {
+            res = reviewDao.deleteById(id);
+            updateSongRating(id);
+        }
+        return res;
     }
 
     @Override
@@ -40,7 +82,24 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public int saveArtistReview(ArtistReview review) {
-        return reviewDao.saveArtistReview(review);
+        int result = reviewDao.saveArtistReview(review);
+        if (result > 0) {
+            updateArtistRating(review.getArtist().getId());
+        }
+        return result;
+    }
+
+    private void updateArtistRating(long artistId) {
+        List<ArtistReview> reviews = reviewDao.findReviewsByArtistId(artistId);
+        float avgRating = (float) reviews.stream().mapToInt(ArtistReview::getRating).average().orElse(0.0);
+        float roundedAvgRating = Math.round(avgRating * 100.0) / 100.0f;
+        int ratingAmount = reviews.size();
+        artistDao.updateRating(artistId, roundedAvgRating, ratingAmount);
+    }
+
+    @Override
+    public boolean hasUserReviewedArtist(long userId, long artistId) {
+        return artistDao.hasUserReviewed(userId, artistId);
     }
 
     @Override
@@ -49,8 +108,40 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
+    public Optional<ArtistReview> findArtistReviewByUserId(long userId, long artistId) {
+        return reviewDao.findArtistReviewByUserId(userId, artistId);
+    }
+
+    @Override
+    public Optional<AlbumReview> findAlbumReviewByUserId(long userId, long albumId) {
+        return reviewDao.findAlbumReviewByUserId(userId, albumId);
+    }
+
+    @Override
+    public Optional<SongReview> findSongReviewByUserId(long userId, long songId) {
+        return reviewDao.findSongReviewByUserId(userId, songId);
+    }
+
+    @Override
     public int saveAlbumReview(AlbumReview review) {
-        return reviewDao.saveAlbumReview(review);
+        int result = reviewDao.saveAlbumReview(review);
+        if (result > 0) {
+            updateAlbumRating(review.getAlbum().getId());
+        }
+        return result;
+    }
+
+    private void updateAlbumRating(long albumId) {
+        List<AlbumReview> reviews = reviewDao.findReviewsByAlbumId(albumId);
+        float avgRating = (float) reviews.stream().mapToInt(AlbumReview::getRating).average().orElse(0.0);
+        float roundedAvgRating = Math.round(avgRating * 100.0) / 100.0f;
+        int ratingAmount = reviews.size();
+        albumDao.updateRating(albumId, roundedAvgRating, ratingAmount);
+    }
+
+    @Override
+    public boolean hasUserReviewedAlbum(long userId, long albumId) {
+        return albumDao.hasUserReviewed(userId, albumId);
     }
 
     @Override
@@ -61,7 +152,24 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public int saveSongReview(SongReview review) {
-        return reviewDao.saveSongReview(review);
+        int result = reviewDao.saveSongReview(review);
+        if (result > 0) {
+            updateSongRating(review.getSong().getId());
+        }
+        return result;
+    }
+
+    private void updateSongRating(long songId) {
+        List<SongReview> reviews = reviewDao.findReviewsBySongId(songId);
+        float avgRating = (float) reviews.stream().mapToInt(SongReview::getRating).average().orElse(0.0);
+        float roundedAvgRating = Math.round(avgRating * 100.0) / 100.0f;
+        int ratingAmount = reviews.size();
+        songDao.updateRating(songId, roundedAvgRating, ratingAmount);
+    }
+
+    @Override
+    public boolean hasUserReviewedSong(long userId, long songId) {
+        return songDao.hasUserReviewed(userId, songId);
     }
 
     @Override
